requirements.txt
PyYAML
jsonschema
cerberus
python-dotenv
requests
urllib3 